Added machine translation docs, isolated translation features to not be utilized unless enabled
