+++ hypha/apply/funds/forms.py
@@ -5,16 +5,15 @@ from operator import methodcaller
 
 import nh3
 from django import forms
+from django.conf import settings
 from django.db.models import Q
 from django.utils.safestring import mark_safe
 from django.utils.translation import gettext_lazy as _
 from wagtail.signal_handlers import disable_reference_index_auto_update
 
 from hypha.apply.categories.models import MetaTerm
-from hypha.apply.translate.utils import get_available_translations
 from hypha.apply.users.models import User
 
-from .fields import LanguageChoiceField
 from .models import (
     ApplicationSubmission,
     AssignedReviewers,
@@ -25,6 +24,9 @@ from .permissions import can_change_external_reviewers
 from .utils import model_form_initial, render_icon
 from .widgets import MetaTermSelect2Widget, Select2MultiCheckboxesWidget
 from .workflow import get_action_mapping
+
+if settings.APPLICATION_TRANSLATIONS_ENABLED:
+    pass
 
 
 class ApplicationSubmissionModelForm(forms.ModelForm):
@@ -377,32 +379,6 @@ class BatchUpdateReviewersForm(forms.Form):
             if not submission.stage.has_external_review:
                 return True
         return False
-
-
-class TranslateSubmissionForm(forms.Form):
-    available_packages = get_available_translations()
-
-    from_lang = LanguageChoiceField("from", available_packages)
-    to_lang = LanguageChoiceField("to", available_packages)
-
-    def clean(self):
-        form_data = self.cleaned_data
-        try:
-            from_code = form_data["from_lang"]
-            to_code = form_data["to_lang"]
-
-            to_packages = get_available_translations([from_code])
-
-            if to_code not in [package.to_code for package in to_packages]:
-                self.add_error(
-                    "to_lang",
-                    "The specified language is either invalid or not installed",
-                )
-
-            return form_data
-        except KeyError as err:
-            # If one of the fields could not be parsed, there is likely bad input being given
-            raise forms.ValidationError("Invalid input selected") from err
 
 
 def make_role_reviewer_fields():
