+++ hypha/apply/funds/tests/test_tags.py
@@ -28,7 +28,7 @@ class TestTemplateTags(TestCase):
             == f'Lorem ipsum dolor <a href="{submission.get_absolute_url()}">{submission.title} <span class="text-gray-400">#{submission.public_id or submission.id}</span></a> sit amet.'
         )
 
-    @override_settings(SUBMISSION_TRANSLATIONS_ENABLED=True)
+    @override_settings(APPLICATION_TRANSLATIONS_ENABLED=True)
     def test_translate_tags_as_applicant(self):
         submission = ApplicationSubmissionFactory()
         request = RequestFactory().get(submission.get_absolute_url())
@@ -41,7 +41,7 @@ class TestTemplateTags(TestCase):
 
         self.assertEqual(output, "")
 
-    @override_settings(SUBMISSION_TRANSLATIONS_ENABLED=True)
+    @override_settings(APPLICATION_TRANSLATIONS_ENABLED=True)
     def test_translate_tags_as_staff(self):
         submission = ApplicationSubmissionFactory()
         request = RequestFactory().get(submission.get_absolute_url())
@@ -54,7 +54,7 @@ class TestTemplateTags(TestCase):
 
         self.assertEqual(output, "<p>some translation stuff</p>")
 
-    @override_settings(SUBMISSION_TRANSLATIONS_ENABLED=False)
+    @override_settings(APPLICATION_TRANSLATIONS_ENABLED=False)
     def test_translate_tags_disabled(self):
         submission = ApplicationSubmissionFactory()
         request = RequestFactory().get(submission.get_absolute_url())
