+++ hypha/apply/funds/tests/test_views.py
@@ -673,7 +673,7 @@ class TestStaffSubmissionView(BaseSubmissionViewTestCase):
         response = SubmissionDetailView.as_view()(request, pk=submission.pk)
         self.assertEqual(response.status_code, 200)
 
-    @override_settings(SUBMISSION_TRANSLATIONS_ENABLED=True)
+    @override_settings(APPLICATION_TRANSLATIONS_ENABLED=True)
     def test_staff_can_see_translate_primary_action(self):
         def assert_view_translate_displayed(submission):
             response = self.get_page(submission)
@@ -690,7 +690,7 @@ class TestStaffSubmissionView(BaseSubmissionViewTestCase):
 
         assert_view_translate_displayed(self.submission)
 
-    @override_settings(SUBMISSION_TRANSLATIONS_ENABLED=False)
+    @override_settings(APPLICATION_TRANSLATIONS_ENABLED=False)
     def test_staff_cant_see_translate_primary_action(self):
         def assert_view_translate_displayed(submission):
             response = self.get_page(submission)
