+++ hypha/apply/translate/utils.py
@@ -1,13 +1,19 @@
+import json
+import re
 from typing import List, Optional, Tuple
 from urllib.parse import parse_qs, urlparse
 
-from argostranslate import package, translate
+import argostranslate
+from bs4 import BeautifulSoup, element
+from django.conf import settings
 from django.http import HttpRequest
+
+from . import translate
 
 
 def get_available_translations(
     from_codes: Optional[List[str]] = None,
-) -> List[package.Package]:
+) -> List[argostranslate.package.Package]:
     """Get languages available for translation
 
     Args:
@@ -17,7 +23,7 @@ def get_available_translations(
         A list of argostranslate package objects that are installed and available.
     """
 
-    available_packages = package.get_installed_packages()
+    available_packages = argostranslate.package.get_installed_packages()
 
     if not from_codes:
         return available_packages
@@ -59,6 +65,170 @@ def get_translation_params(
 
 def get_lang_name(code: str) -> str | None:
     try:
-        return translate.get_language_from_code(code).name
+        return argostranslate.translate.get_language_from_code(code).name
     except AttributeError:
         return None
+
+
+def get_language_choices_json(request: HttpRequest) -> str:
+    """Generate a JSON output of available translation options
+
+    Utilized for populating the reactive form fields on the client side
+
+    Args:
+        request: an `HttpRequest` containing an "Hx-Current-Url" header to extract current translation params from
+
+    Returns:
+        A JSON string in the format of:
+
+        ```
+        [
+            {
+                "value": "<from language code>",
+                "label": "<from language name>",
+                "to": [
+                    {
+                        "value": "<to language code>",
+                        "label": "<from language name>"
+                        "selected": <bool if selected by default>
+                    }
+                ],
+                "selected": <bool if selected by default>
+            },
+            ...
+        ]
+        ```
+    """
+    available_translations = get_available_translations()
+    from_langs = {package.from_code for package in available_translations}
+    default_to_lang = settings.LANGUAGE_CODE if settings.LANGUAGE_CODE else None
+    default_from_lang = None
+
+    # If there's existing lang params, use those as the default in the form
+    # ie. the user has an active translation for ar -> en, those should be selected in the form
+    if (current_url := request.headers.get("Hx-Current-Url")) and (
+        params := get_translation_params(current_url)
+    ):
+        default_from_lang, default_to_lang = params
+
+    choices = []
+    for lang in from_langs:
+        to_langs = [
+            package.to_code
+            for package in available_translations
+            if package.from_code == lang
+        ]
+
+        # Set the default selection to be the default_to_lang if it exists in the to_langs list,
+        # otherwise use the first value in the list.
+        selected_to = (
+            default_to_lang
+            if default_to_lang and default_to_lang in to_langs
+            else to_langs[0]
+        )
+
+        to_choices = [
+            {
+                "value": to_lang,
+                "label": get_lang_name(to_lang),
+                "selected": to_lang == selected_to,
+            }
+            for to_lang in to_langs
+        ]
+
+        choices.append(
+            {
+                "value": lang,
+                "label": get_lang_name(lang),
+                "to": to_choices,
+                "selected": lang == default_from_lang,
+            }
+        )
+
+    return json.dumps(choices)
+
+
+def translate_application_form_data(application, from_code: str, to_code: str) -> dict:
+    """Translate the content of an application's live revision `form_data`.
+    Will parse fields that contain both plaintext & HTML, extracting & replacing strings.
+
+    NOTE: Mixed formatting like `<p>Hey from <strong>Hypha</strong></p>` will result in a
+    string that is stripped of text formatting (untranslated: `<p>Hey from Hypha</p>`). On
+    the other hand, unmixed strings like `<p><strong>Hey from Hypha</strong></p>` will be
+    replaced within formatting tags.
+
+    Args:
+        application: the  application to translate
+        from_code: the ISO 639 code of the original language
+        to_code: the ISO 639 code of the language to translate to
+
+    Returns:
+        The `form_data` with values translated (including nested HTML strings)
+
+    Raises:
+        ValueError if an invalid `from_code` or `to_code` is requested
+    """
+    form_data: dict = application.live_revision.form_data
+
+    translated_form_data = form_data.copy()
+
+    # Only translate content fields or the title - don't with name, email, etc.
+    translated_form_data["title"] = translate.translate(
+        form_data["title"], from_code, to_code
+    )
+
+    # RegEx to match wagtail's generated field UIDs - ie. "97c51cea-ab47-4a64-a64a-15d893788ef2"
+    uid_regex = re.compile(r"([a-z]|\d){8}(-([a-z]|\d){4}){3}-([a-z]|\d){12}")
+    fields_to_translate = [
+        key
+        for key in form_data
+        if uid_regex.match(key) and isinstance(form_data[key], str)
+    ]
+
+    for key in fields_to_translate:
+        field_html = BeautifulSoup(form_data[key], "html.parser")
+        if field_html.find():  # Check if BS detected any HTML
+            for field in field_html.find_all(has_valid_str):
+                # Removes formatting if mixed into the tag to prioritize context in translation
+                # ie. `<p>Hey <strong>y'all</strong></p>` -> `<p>Hey y'all</p>` (but translated)
+                to_translate = field.string if field.string else field.text
+                field.clear()
+                field.string = translate.translate(to_translate, from_code, to_code)
+
+            translated_form_data[key] = str(field_html)
+        # Ensure the field value isn't empty & translate as is
+        elif form_data[key].strip():
+            translated_form_data[key] = translate.translate(
+                form_data[key], from_code, to_code
+            )
+
+    return translated_form_data
+
+
+def has_valid_str(tag: element.Tag) -> bool:
+    """Checks that an Tag contains a valid text element and/or string.
+
+    Args:
+        tag: a `bs4.element.Tag`
+    Returns:
+        bool: True if has a valid string that isn't whitespace or `-`
+    """
+    text_elem = tag.name in ["span", "p", "strong", "em", "td", "a"]
+
+    try:
+        # try block logic handles elements that have text directly in them
+        # ie. `<p>test</p>` or `<em>yeet!</em>` would return true as string values would be contained in tag.string
+        ret = bool(
+            text_elem
+            and tag.find(string=True, recursive=False)
+            and tag.string.strip(" -\n")
+        )
+        return ret
+    except AttributeError:
+        # except block logic handles embedded tag strings where tag.string == None but the specified tag DOES contain a string
+        # ie. `<p>Hypha is <strong>cool</strong></p>` contains the string "Hypha is" but due to the strong tag being mixed in will
+        # have None for the tag.string value.
+        # tags like `<p> <em>Hypha rocks</em> </p>` will return false as the <p> tag contains no valid strings, it's child does.
+        tag_contents = "".join(tag.find_all(string=True, recursive=False))
+        ret = bool(tag.text and tag.text.strip() and tag_contents.strip())
+        return ret
