+++ requirements.txt
@@ -3,7 +3,6 @@ scout-apm==3.1.0
 sentry-sdk==2.8.0
 
 # Production dependencies
-argostranslate==1.9.6
 Babel==2.14.0
 boto3==1.34.66
 celery==5.3.6
